backend/python/requirements.txt
pytesseract>=0.3.10
pandas>=2.0.0
PyMuPDF>=1.23.0
pdf2image>=1.16.3
Pillow>=10.0.0
